Refactored LoHaModel and LoKrModel
